API/requirements.txt
Babel==2.4.0
Django==1.11.2
django-filter==1.0.4
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
olefile==0.44
phonenumberslite==8.6.0
Pillow==4.2.1
psycopg2==2.7.1
pytz==2017.2
